mmu: fix integer promotion issue on rv32
Without this change a uint32_t overflow would occur resulting in an
incorrect ptbase address and causing SV32 to not work correctly.

+++ vp/src/core/common/mmu.h
@@ -126,7 +126,7 @@ struct GenericMMU {
 
     vm_info decode_vm_info(PrivilegeLevel prv) {
         assert(prv <= SupervisorMode);
-        uint64_t ptbase = core.csrs.satp.ppn << PGSHIFT;
+        uint64_t ptbase = (uint64_t)core.csrs.satp.ppn << PGSHIFT;
         unsigned mode = core.csrs.satp.mode;
         switch (mode) {
             case SATP_MODE_SV32:
